Edited the following lines for grammar & precision
Line 25. Used possession syntax.
Line 76. Added clarity and brevity.
Line 101. Changed 'above tag' to 'component'.
Line 111.  Added comma, brevity and made it overall more concise. Ex "recognizable when compared to JSX".
..cont. I added this at the end of Line 111. No idea if it's accurate and therefore appropriate.'.....however Riot does allow that HTML and JavaScript can be mixed within expressions.'.

Line 113. stuff is now 'code'.
Line 115. "Just less OF everything with Riot". added 'of' and 'with Riot'.
Line 117. "We believe" not we think.
Line 128. Removed some words and tweaked. PS. "Given the enormous challenge.... " This sentence seems meaningless. Perhaps there is an intended meaning and if so I'd tweak it or remove it.
Line 130. diff'ing. You are referring to diff correct? Then ' is needed then to ensures it's understood without the mental requirement of linking the intent and context to understand it's meaning.
Line 132. Tweaked. Last sentence should or could potentially be removed as it appears excessively verbose or perhaps blended into the second last sentence that relates to it.

Line 137. Shortened and combined with Lin 138.
Line 139.  Riot 'also' comes..... also relates to the preview paragraph and assures the reader of there is equivalence in this area you just praised.
Line 141. brevity again and showed that Riot replaces both Flux and React which isn't clear.
Line 141. 'it's just a library to help you achieve 'the' goal'. What goal? Can we be more specific here? It's a moment to communicate it's value. FYI. All of line 141 confuses me. I'm sure once I know more about Riot I'll get the intent here, but as a newb to the project I felt there was some back and forth. Is Riot like React and Flux? Just like React? This is exacerbated by Line 143's link to RiotControl.

Line 143. changed observable to event emitter for consistency.

Line 171, 176 178. Brevity and precision again. 
Line 180-186. Is this what Polymer does? I would assume so and how is Riot different? Very vague here. Like the writer was tired at this point. I can't help with this part yet, since I'm too new to the project to comment on technical components specific to Riot.

Line 204. Rewrote everything.